requirements.txt
botocore>=1.34.0
boto3>=1.34.0
kafka-python>=2.0.2
pyosmium>=3.7.0
psycopg2-binary>=2.9.9
requests>=2.32.3
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
shapely>=2.0.4
pyproj>=3.6.1
scikit-learn>=1.5.1
numpy>=1.26.4
